[atlas-gradle-plugin] 不在resolvedependency的时候解压ap，打乱任务流程
