+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tasks/PrepareAPTask.java
@@ -344,9 +344,6 @@ public class PrepareAPTask extends DefaultTask {
             apContext.setApExploredFolder(explodedDir);
             prepareAPTask.setApBaseFile(apBaseFile);
             prepareAPTask.setExplodedDir(explodedDir);
-            if (variantContext.getAtlasExtension().getTBuildConfig().isIncremental()) {
-                prepareAPTask.setDidWork(false);
-            }
         }
     }
 }
