+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tasks/incremental/ApDependencies.java
@@ -211,23 +211,36 @@ package com.taobao.android.builder.tasks.incremental;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.Map;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipFile;
 
 import com.alibaba.fastjson.JSON;
 
+import com.google.common.base.Predicate;
+import com.google.common.collect.Collections2;
+import com.google.common.collect.Iterables;
 import com.google.common.collect.Maps;
 import com.taobao.android.builder.dependency.output.DependencyJson;
 import com.taobao.android.builder.dependency.parser.ResolvedDependencyInfo;
-import org.apache.commons.io.FileUtils;
+import com.taobao.android.builder.extension.TBuildType;
+import org.apache.commons.io.IOUtils;
+import org.gradle.api.Nullable;
 import org.gradle.api.Project;
+import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.artifacts.ModuleIdentifier;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.dsl.DependencyHandler;
 import org.gradle.api.internal.artifacts.DefaultModuleIdentifier;
 import org.gradle.api.internal.artifacts.ivyservice.ivyresolve.strategy.DefaultVersionComparator;
+
+import static com.android.build.gradle.internal.api.ApContext.DEPENDENCIES_FILENAME;
+import static com.google.common.base.Strings.isNullOrEmpty;
 
 /**
  * Created by chenhjohn on 2017/5/1.
@@ -248,16 +261,20 @@ public class ApDependencies /*extends BaseTask*/ {
 
     private final DependencyJson apDependencyJson;
 
-    public ApDependencies(Project project, File baseDependenciesFile) {
+    public ApDependencies(Project project, TBuildType tBuildType) {
         this.dependencies = project.getDependencies();
+        File apBaseFile;
+        apBaseFile = getBaseApFile(project, tBuildType);
 
-        try {
-            apDependencyJson = JSON.parseObject(FileUtils.readFileToString(baseDependenciesFile), DependencyJson.class);
-
+        try (ZipFile zip = new ZipFile(apBaseFile)) {
+            ZipEntry entry = zip.getEntry(DEPENDENCIES_FILENAME);
+            try (InputStream in = zip.getInputStream(entry)) {
+                apDependencyJson = JSON.parseObject(IOUtils.toString(in, StandardCharsets.UTF_8), DependencyJson.class);
+            }
         } catch (IOException e) {
-            throw new RuntimeException(String.format("Unable to read %1$s.", baseDependenciesFile.getAbsolutePath()),
-                                       e);
+            throw new RuntimeException("Unable to read dependencies.txt from " + apBaseFile.getAbsolutePath(), e);
         }
+
         for (String mainDex : apDependencyJson.getMainDex()) {
             addDependency(mainDex, null);
         }
@@ -270,6 +287,31 @@ public class ApDependencies /*extends BaseTask*/ {
             }
         }
 
+    }
+
+    private File getBaseApFile(Project project, TBuildType tBuildType) {
+        File apBaseFile;
+        File buildTypeBaseApFile = tBuildType.getBaseApFile();
+        if (null != buildTypeBaseApFile && buildTypeBaseApFile.exists()) {
+            apBaseFile = buildTypeBaseApFile;
+        } else if (!isNullOrEmpty(tBuildType.getBaseApDependency())) {
+            String apDependency = tBuildType.getBaseApDependency();
+            // Preconditions.checkNotNull(apDependency,
+            //                            "You have to specify the baseApFile property or the baseApDependency
+            // dependency");
+            Dependency dependency = project.getDependencies().create(apDependency);
+            Configuration configuration = project.getConfigurations().detachedConfiguration(dependency);
+            configuration.setTransitive(false);
+            apBaseFile = Iterables.getOnlyElement(Collections2.filter(configuration.getFiles(), new Predicate<File>() {
+                @Override
+                public boolean apply(@Nullable File file) {
+                    return file.getName().endsWith(".ap");
+                }
+            }));
+        } else {
+            throw new IllegalStateException("AP is missing");
+        }
+        return apBaseFile;
     }
 
     public boolean isMainLibrary(ResolvedDependencyInfo dependencyInfo) {
@@ -299,42 +341,4 @@ public class ApDependencies /*extends BaseTask*/ {
     }
 
     private String getVersion(String baseLibrary) {return baseLibrary.substring(baseLibrary.lastIndexOf(':') + 1);}
-
-    // // ----- Config Action -----
-    // public static final class ConfigAction extends MtlBaseTaskAction<ApDependencies> {
-    //     @NonNull
-    //     private final VariantScope scope;
-    //
-    //     public ConfigAction(VariantContext variantContext, BaseVariantOutputData baseVariantOutputData) {
-    //         super(variantContext, baseVariantOutputData);
-    //         this.scope = baseVariantOutputData.getScope().getVariantScope();
-    //     }
-    //
-    //     @Override
-    //     @NonNull
-    //     public String getName() {
-    //         return scope.getTaskName("correct", "AtlasDependencies");
-    //     }
-    //
-    //     @Override
-    //     @NonNull
-    //     public Class<ApDependencies> getType() {
-    //         return ApDependencies.class;
-    //     }
-    //
-    //     @Override
-    //     public void execute(@NonNull ApDependencies correctDependencies) {
-    //         BaseVariantData<? extends BaseVariantOutputData> variantData = scope.getVariantData();
-    //
-    //         final GradleVariantConfiguration variantConfiguration = variantData.getVariantConfiguration();
-    //         // correctDependencies.setAndroidBuilder(scope.getGlobalScope().getAndroidBuilder());
-    //         // correctDependencies.setVariantName(scope.getVariantConfiguration().getFullName());
-    //         // correctDependencies.atlasDependencyTree = AtlasBuildContext.androidDependencyTrees.get(
-    //         //     correctDependencies.getVariantName());
-    //         //
-    //         // ConventionMappingHelper.map(correctDependencies, "baseDependenciesFile", () -> {
-    //         //     return variantContext.apContext.getBaseDependenciesFile();
-    //         // });
-    //     }
-    // }
 }
